+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - avoid installing "undefined" npm package when importing authored components
 - add authentication fallback to ssh-key in case the ssh-agent is enabled but failed to authenticate
 - improve authentication error message to clearly indicate the various strategies failures
-- fix `build` command with no cache flag 
+- reintroduce `-c` alias for `--no-cache` flag in `bit build` command 
 
 ## [14.0.5] - 2019-04-07