+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- fix symlink to binary (or unsupported) files dependencies when installed via npm and have dists
+
 ## [14.0.5] - 2019-04-07
 
 - fix `remove` command to not delete dependencies files from the scope as they might belong to other components